+++ spec/ruby/core/ractor/port/append_spec.rb
@@ -1,0 +1,8 @@
+require_relative '../../../spec_helper'
+require_relative 'shared/send'
+
+ruby_version_is "4.0" do
+  describe "Ractor::Port#<<" do
+    it_behaves_like :ractor_port_send, :<<
+  end
+end